pyproject.toml
[project]
name = "lending-club-risk"
version = "0.1.0"
description = "Credit Risk Management Pipeline: ML + Conformal Prediction + Operations Research"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
authors = [{ name = "Carlos Vergara" }]

dependencies = [
    # ── Core Data ──
    "pandas>=2.2,<3",
    "polars>=1.20",
    "numpy>=1.26,<3",
    "pyarrow>=18.0",
    # ── ML Core ──
    "scikit-learn>=1.6,<2",
    "catboost>=1.2",
    "lightgbm>=4.5",
    "optuna>=4.1",
    "shap>=0.47",
    # ── Conformal Prediction ──
    "mapie>=1.3.0,<2",
    "crepes>=0.7",
    # ── Time Series (Nixtla) ──
    "statsforecast>=2.0",
    "mlforecast>=0.13",
    "hierarchicalforecast>=1.0",
    # ── Survival Analysis ──
    "lifelines>=0.30",
    "scikit-survival>=0.24",
    # ── Causal ML ──
    "econml>=0.16",
    "dowhy>=0.12",
    # ── Optimization ──
    "pyomo>=6.8",
    "highspy>=1.10",
    "cvxpy>=1.6",
    # ── Feature Engineering ──
    "optbinning>=0.19",
    # ── MLOps ──
    "dvc>=3.56",
    "mlflow>=2.20",
    "dagshub>=0.4",
    "pandera>=0.22",
    # ── API & Dashboard ──
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "streamlit>=1.42",
    "pydantic>=2.10",
    "httpx>=0.27",
    # ── Data Platform ──
    "duckdb>=1.1",
    # ── MCP (Model Context Protocol) ──
    "mcp[cli]>=1.2",
    # ── Visualization ──
    "matplotlib>=3.9",
    "seaborn>=0.13",
    "plotly>=5.24",
    # ── Utilities ──
    "joblib>=1.4",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "tqdm>=4.67",
    "loguru>=0.7",
    "streamlit-flow-component>=1.6.1",
    "streamlit-mermaid>=0.3.0",
    "dvc-gdrive>=3.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=6.0",
    "ruff>=0.9",
    "nbstripout>=0.8",
    "jupytext>=1.17",
    "ipykernel>=6.29",
    "jupyter>=1.1",
    "pre-commit>=4.0",
]
spo = [
    "pyepo>=0.5",
]
platform = [
    "dbt-duckdb>=1.10",
    "feast>=0.59",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# ── Ruff ──
[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

# ── Pytest ──
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests",
]

# ── Jupytext ──
[tool.jupytext]
formats = "ipynb,py:percent"
